Add common base condition class for JSON Api and Rpc for common verify method.
